Change the ...ImageTag and ...Image interaction to be more apparent
This is indirect to preserve as much of the previous behavior as
possible, the only change should be when an image is overridden(but
without a tag inline) and
the imageTag is also overriden. Previously this would result in an
implicit 'latest' tag being used, with this change the tag specified
in the imageTag would be used.